pyproject.toml
[project]
name = "technical-interview"
version = "0.1.0"
description = "ML package as part of a technical interview"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grid2op==1.12.1",
    "lightsim2grid==0.10.3",
    "numba>=0.62.1",
    "numpy==2.3.3",
    "polars==1.33.1",
    "pyarrow==21.0.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "structlog>=25.5.0",
    "tqdm==4.67.1",
    "typer>=0.20.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "nbformat>=5.10.4",
    "plotly>=6.4.0",
    "pytest>=8.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
